requirements.txt
python-telegram-bot==22.0
flask
gunicorn
torch
kagglehub